+++ netbox/templates/dcim/device.html
@@ -220,7 +220,7 @@
                     None found
                 </div>
             {% endif %}
-            {% if perms.ipam.add_ipaddress %}
+            {% if perms.dcim.add_serviceport %}
                 <div class="panel-footer text-right">
                     <a href="{% url 'dcim:serviceport_add' pk=device.pk %}" class="btn btn-xs btn-primary">
                         <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>
